Fixed conditional and directory issues
Fixed the issue where testing for true is not the same as testing
for yes in build.sh at the end of the script when deciding which
deploy script to run. Previously it would just exit without
running either.

Fixed the directory issue in deploy-rpc.sh where it would copy
openstack_deploy into /etc/openstack_deploy/openstack_deploy
instead of /etc/openstack_deploy where it belongs.